requirements.txt
discord.py[voice]>=1.7.3
pynacl>=1.5.0
python-dotenv>=0.19.2
requests==2.28.1
emoji==2.0.0
redis==4.3.4
yt-dlp==2022.7.18
pytz>=2021.3
